Extract all sorts into a main compoonent so we can bind the controller
