+++ app/components/queries/sort_by_component.rb
@@ -29,21 +29,25 @@
 # ++
 
 class Queries::SortByComponent < ApplicationComponent
-  options :order,
-          :available_orders
+  options :query
 
-  def select_options
-    options_for_select(
-      available_orders.map { |order| [order[:name], order[:id]] },
-      order&.attribute
-    )
+  def current_orders
+    JSON.dump(query.orders.map { |order| [order.attribute, order.direction] })
   end
 
-  def order_asc?
-    order&.direction == :asc
+  def order_limit
+    3
   end
 
-  def order_desc?
-    order&.direction == :desc
+  def available_orders
+    @available_orders ||= begin
+      all_selectable_columns = helpers.projects_columns_options
+      all_order_keys = ::Queries::Register.orders[query.class]&.map(&:key)
+
+      # Keys from the order can be symbols, strings or regexes
+      all_selectable_columns.select do |column_option|
+        all_order_keys.any? { |order_key| order_key === column_option[:id] }
+      end
+    end
   end
 end
